+++ src/Enum/Subscription/SubscriptionProductId.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace App\Enum\Subscription;
+
+enum SubscriptionProductId: string
+{
+  case SELEN_PREMIUM = 'selen_premium';
+}